+++ src/plone/app/jsonfield/helpers.py
@@ -1,4 +1,6 @@
 # _*_ coding: utf-8 _*_
+from .compat import EMPTY_STRING
+from .compat import NO_VALUE
 from plone.app.jsonfield.compat import _
 from plone.app.jsonfield.compat import json
 from zope.interface import Invalid
@@ -12,6 +14,9 @@ __author__ = 'Md Nazrul Islam<email2nazrul@gmail.com>'
 
 def parse_json_str(str_val, encoding='utf-8'):
     """ """
+    if str_val in (NO_VALUE, EMPTY_STRING, None):
+        # No parsing for empty value
+        return None
     try:
         json_dict = json.loads(str_val, encoding=encoding)
     except ValueError as exc: