+++ src/plone/app/jsonfield/tests/test_helpers.py
@@ -1,5 +1,6 @@
 # _*_ coding: utf-8 _*_
 from . import FHIR_FIXTURE_PATH
+from plone.app.jsonfield import compat
 from plone.app.jsonfield import helpers
 from plone.app.jsonfield.testing import PLONE_APP_JSON_FIELD_INTEGRATION_TESTING
 from zope.interface import Invalid
@@ -32,3 +33,14 @@ class HelperIntegrationTest(unittest.TestCase):
             raise AssertionError('Code shouldn\'t come here! as invalid json string is provided')
         except Invalid:
             pass
+
+    def test_parse_json_str_with_empty(self):
+        """ """
+        value = helpers.parse_json_str(compat.NO_VALUE)
+        self.assertIsNone(value)
+
+        value = helpers.parse_json_str(compat.EMPTY_STRING)
+        self.assertIsNone(value)
+
+        value = helpers.parse_json_str(None)
+        self.assertIsNone(value)